refactor: simulator to runner
