+++ src/orquestra/integrations/cirq/runner/__init__.py
@@ -1,7 +1,7 @@
 ################################################################################
 # © Copyright 2021-2022 Zapata Computing Inc.
 ################################################################################
-from .simulator import CirqSimulator
+from .cirq_simulator import CirqSimulator
 
 try:
     from .qsim_simulator import QSimSimulator
